multi-hop-reasoning/requirements.txt
# File: requirements.txt
groq
sentence-transformers
faiss-cpu
pandas
numpy
tqdm
scikit-learn
python-dotenv